requirements.txt
pandas
requests
scikit-learn
python-dotenv
google-genai
